Added RSS Feed workflow
Added RSS Feed loader workflow for n8n. Takes a RSS feed, such as from rss-bridge.org, extracts the URLs from the last day, clears a Google doc and loads them with the URLs. This can be used with notebookLM to have a single doc loaded in a notebook that you can refresh and get information on the latest articles